+++ ftx-sivan-shared/src/index.ts
@@ -1,2 +1,2 @@
-export * from './lib/ftx-sivan-shared/ftx-sivan-shared.component';
 export * from './lib/directive';
+export * from './lib/util';
