--- ftx-sivan-shared/src/lib/ftx-sivan-shared/ftx-sivan-shared.component.html
@@ -1,1 +1,0 @@
-<p>ftx-sivan-shared works!</p>
