--- ftx-sivan-shared/src/lib/ftx-sivan-shared/ftx-sivan-shared.component.ts
@@ -1,12 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'coagulu-ftx-sivan-shared',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './ftx-sivan-shared.component.html',
-  styleUrls: ['./ftx-sivan-shared.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class FtxSivanSharedComponent {}
