+++ ftx-sivan-shared/src/lib/util/index.ts
@@ -1,0 +1,1 @@
+export * from './injectControl';
